cleaning up names

+++ src/serializers.pxd
@@ -1,17 +1,17 @@
 from .variant cimport *
 
-ctypedef variant_container (*serializer_t)(const unsigned char *, long *)
+ctypedef variant_container (*serializer_t)(const unsigned char *, uint32_t *)
 
-cdef variant_container get_int8(const unsigned char * binbuffer, long *offset)
-cdef variant_container get_uint8(const unsigned char * binbuffer, long *offset)
-cdef variant_container get_int16(const unsigned char * binbuffer, long *offset)
-cdef variant_container get_uint16(const unsigned char * binbuffer, long *offset)
-cdef variant_container get_int32(const unsigned char * binbuffer, long *offset)
-cdef variant_container get_uint32(const unsigned char * binbuffer, long *offset)
-cdef variant_container get_int64(const unsigned char * binbuffer, long *offset)
-cdef variant_container get_uint64(const unsigned char * binbuffer, long *offset)
-cdef variant_container get_double(const unsigned char * binbuffer, long *offset)
-cdef variant_container get_float(const unsigned char * binbuffer, long *offset)
+cdef variant_container get_int8(const unsigned char * binbuffer, uint32_t *offset)
+cdef variant_container get_uint8(const unsigned char * binbuffer, uint32_t *offset)
+cdef variant_container get_int16(const unsigned char * binbuffer, uint32_t *offset)
+cdef variant_container get_uint16(const unsigned char * binbuffer, uint32_t *offset)
+cdef variant_container get_int32(const unsigned char * binbuffer, uint32_t *offset)
+cdef variant_container get_uint32(const unsigned char * binbuffer, uint32_t *offset)
+cdef variant_container get_int64(const unsigned char * binbuffer, uint32_t *offset)
+cdef variant_container get_uint64(const unsigned char * binbuffer, uint32_t *offset)
+cdef variant_container get_double(const unsigned char * binbuffer, uint32_t *offset)
+cdef variant_container get_float(const unsigned char * binbuffer, uint32_t *offset)
 
 ctypedef int (*deserializer_t)(bytearray, object) except -1
